gateway/app/src/requirements.txt
fastapi
pydantic
uvicorn
requests==2.31.0
PyYAML==6.0.1
